+++ src/main/java/com/webhook/receiver/slack/exception/PropertyNotFoundException.java
@@ -1,0 +1,7 @@
+package com.webhook.receiver.slack.exception;
+
+public class PropertyNotFoundException extends RuntimeException {
+    public PropertyNotFoundException(String message) {
+        super(message);
+    }
+}
