+++ src/main/java/com/webhook/receiver/slack/sender/Notifier.java
@@ -1,0 +1,7 @@
+package com.webhook.receiver.slack.sender;
+
+import com.webhook.receiver.slack.vo.WebhookPayload;
+
+public interface Notifier {
+    boolean send(WebhookPayload webhookPayload);
+}
