+++ src/main/java/com/webhook/receiver/slack/vo/AlarmChecker.java
@@ -1,0 +1,19 @@
+package com.webhook.receiver.slack.vo;
+
+
+public abstract class AlarmChecker<T> extends Checker {
+    protected T detectedValue;
+    
+    public AlarmChecker(T detectedValue) {
+        this.detectedValue = detectedValue;
+    }
+    
+    public T getDetectedValue() {
+        return detectedValue;
+    }
+    
+    @Override
+    public String getDetectedValueString() {
+        return detectedValue.toString();
+    }
+}
