+++ src/main/java/com/webhook/receiver/slack/vo/BooleanValueAgentChecker.java
@@ -1,0 +1,15 @@
+package com.webhook.receiver.slack.vo;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonTypeName;
+
+import java.util.List;
+
+@JsonTypeName("BooleanValueAgentChecker")
+public class BooleanValueAgentChecker extends AgentChecker<Boolean> {
+    
+    @JsonCreator
+    public BooleanValueAgentChecker(List<DetectedAgent<Boolean>> detectedValue) {
+        super(detectedValue);
+    }
+}
