+++ src/main/java/com/webhook/receiver/slack/vo/DetectedAgent.java
@@ -1,0 +1,22 @@
+package com.webhook.receiver.slack.vo;
+
+public class DetectedAgent<T> {
+    private String agentId;
+    private T agentValue;
+    
+    public String getAgentId() {
+        return agentId;
+    }
+    
+    public T getAgentValue() {
+        return agentValue;
+    }
+    
+    @Override
+    public String toString() {
+        return "{" +
+                "agentId='" + agentId + '\'' +
+                ", agentValue=" + agentValue +
+                '}';
+    }
+}
