+++ src/main/java/com/webhook/receiver/slack/vo/UserGroup.java
@@ -1,0 +1,22 @@
+package com.webhook.receiver.slack.vo;
+
+import java.util.List;
+
+public class UserGroup {
+    
+    private String userGroupId;
+    private List<UserMember> userGroupMembers;
+    
+    public UserGroup(String userGroupId, List<UserMember> userGroupMembers) {
+        this.userGroupId = userGroupId;
+        this.userGroupMembers = userGroupMembers;
+    }
+    
+    public String getUserGroupId() {
+        return userGroupId;
+    }
+    
+    public List<UserMember> getUserGroupMembers() {
+        return userGroupMembers;
+    }
+}
