+++ src/main/java/com/webhook/receiver/slack/vo/sender/SlackPayload.java
@@ -1,0 +1,25 @@
+package com.webhook.receiver.slack.vo.sender;
+
+import java.util.List;
+
+public class SlackPayload {
+    private String text;
+    private List<SlackAttachment> attachments;
+    
+    public SlackPayload(String text, List<SlackAttachment> attachments) {
+        this.text = text;
+        this.attachments = attachments;
+    }
+    
+    public SlackPayload(List<SlackAttachment> attachments) {
+        this.attachments = attachments;
+    }
+    
+    public String getText() {
+        return text;
+    }
+    
+    public List<SlackAttachment> getAttachments() {
+        return attachments;
+    }
+}
