--- src/main/java/com/webhook/receiver/slack/webhook/config/WebConfiguration.java
@@ -1,14 +1,0 @@
-package com.webhook.receiver.slack.webhook.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class WebConfiguration {
-    
-    @Bean
-    RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
-}
