--- src/main/java/com/webhook/receiver/slack/webhook/exception/PropertyNotFoundException.java
@@ -1,7 +1,0 @@
-package com.webhook.receiver.slack.webhook.exception;
-
-public class PropertyNotFoundException extends RuntimeException {
-    public PropertyNotFoundException(String message) {
-        super(message);
-    }
-}
