--- src/main/java/com/webhook/receiver/slack/webhook/sender/Notifier.java
@@ -1,7 +1,0 @@
-package com.webhook.receiver.slack.webhook.sender;
-
-import com.webhook.receiver.slack.webhook.vo.WebhookPayload;
-
-public interface Notifier {
-    boolean send(WebhookPayload webhookPayload);
-}
