--- src/main/java/com/webhook/receiver/slack/webhook/vo/AlarmChecker.java
@@ -1,19 +1,0 @@
-package com.webhook.receiver.slack.webhook.vo;
-
-
-public abstract class AlarmChecker<T> extends Checker {
-    protected T detectedValue;
-    
-    public AlarmChecker(T detectedValue) {
-        this.detectedValue = detectedValue;
-    }
-    
-    public T getDetectedValue() {
-        return detectedValue;
-    }
-    
-    @Override
-    public String getDetectedValueString() {
-        return detectedValue.toString();
-    }
-}
