--- src/main/java/com/webhook/receiver/slack/webhook/vo/DataSourceAlarmListValueAgentChecker.java
@@ -1,15 +1,0 @@
-package com.webhook.receiver.slack.webhook.vo;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonTypeName;
-
-import java.util.List;
-
-@JsonTypeName("DataSourceAlarmListValueAgentChecker")
-public class DataSourceAlarmListValueAgentChecker extends AgentChecker<DataSourceAlarmList> {
-    
-    @JsonCreator
-    public DataSourceAlarmListValueAgentChecker(List<DetectedAgent<DataSourceAlarmList>> detectedValue) {
-        super(detectedValue);
-    }
-}
