+++ src/main/resources/application.properties
@@ -1,3 +1,3 @@
 server.port=8085
 
-slack.webhook.url=
+slack.incoming.webhook.url=
